Length of a trachea from the infiltration point to the nearest TLC
My Neuroscience lab is using Python to analyze the trachea anatomy. In this program, my goal is finding the length from infiltration point to the nearest TLC. The code is originally made in Jupiter Notebook. However, in order not to publish the private information from the lab, I converted them into a Python file. After the paper is published, I will add more information (neuroscience concepts and how the data look like) for better explanation.